# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Sponsor:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo_url:
        type: string
      name:
        type: string
      order_number:
        type: integer
      updated_at:
        type: string
      visible:
        type: boolean
      website_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dekoninklijkeloop.nl
    name: API Support
    url: http://www.dekoninklijkeloop.nl
  description: API voor de De Koninklijke Loop email service - inclusief registratie,
    deelnemersbeheer, CMS en meer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DKL Email Service API
  version: "1.0"
paths:
...
  /api/sponsors:
    get:
      consumes:
      - application/json
      description: Returns all visible sponsors ordered by order_number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sponsor'
            type: array
      summary: Get visible sponsors
      tags:
      - Sponsors
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Creates a new sponsor. Supports both JSON and multipart/form-data
        for logo upload.
      parameters:
      - description: Sponsor data (JSON)
        in: body
        name: sponsor
        required: true
        schema:
          $ref: '#/definitions/models.Sponsor'
      - description: Logo image file
        in: formData
        name: logo
        type: file
      - description: Sponsor name (multipart)
        in: formData
        name: name
        type: string
      - description: Sponsor description (multipart)
        in: formData
        name: description
        type: string
      - description: Website URL (multipart)
        in: formData
        name: website_url
        type: string
      - description: Order number (multipart)
        in: formData
        name: order_number
        type: integer
      - description: Is active (multipart)
        in: formData
        name: is_active
        type: boolean
      - description: Is visible (multipart)
        in: formData
        name: visible
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Sponsor'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create sponsor
      tags:
      - Sponsors
  /api/sponsors/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a sponsor
      parameters:
      - description: Sponsor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete sponsor
      tags:
      - Sponsors
    get:
      consumes:
      - application/json
      description: Returns a specific sponsor by ID
      parameters:
      - description: Sponsor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sponsor'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get sponsor by ID
      tags:
      - Sponsors
    put:
      consumes:
      - application/json
      description: Updates an existing sponsor
      parameters:
      - description: Sponsor ID
        in: path
        name: id
        required: true
        type: string
      - description: Sponsor data
        in: body
        name: sponsor
        required: true
        schema:
          $ref: '#/definitions/models.Sponsor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sponsor'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update sponsor
      tags:
      - Sponsors
  /api/sponsors/admin:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of all sponsors
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sponsor'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all sponsors
      tags:
      - Sponsors
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token voor geauthenticeerde endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"